CSV creates and appends EMG data
The emg data is stored in a CSV called EMG_data.csv. If the file doesn't exist, it will create a new one and add headers. If it exists, it will append to it.

After each trial, the csv file should be renamed and/or moved to another folder so it is not overwritten.