New Moving Avg Filter calculations
Moving average filter implemented in the LineChartController_EMG02.cs

This file doesn't read and writes the average moving EMG from a csv. However, it still requires the raw EMG to be read from a CSV.

The average moving filter contains the sum of the Math.abs( ) values of the last readings.