EMG Pod 08 issue fixed + CSVs save
The Pod 08 EMG data can now be accessed. It seemed like emg[ ] size is 7 in the first loop iterations, but becomes 8 as soon as the Myoband connects with Unity. Not sure why this happens, but the problem is fixed through a counter loop which skips the first readings.

Datapoint storing and saving frequency is 14.3Hz. (Smoothed readings every 70ms).

CSV timestamps are save correctly too in both CSV files.

NOTE: There is still an issue with EMG Pod 01 and 06 always having +1 more element in the PRC arrays. Current code removes the last reading to prevent any issues.