Arm and Pear icons added
Arm and pear icons added.

The 2D arm icon represents the rotation of the patient. The 2D pear icon represents the target. The patient should "CATCH THE PEAR". This helps regulate the speed the task is performed at, further helping to standarise the motion.

Also, fill radial plot color now changes depending on the error amount