Progress Bar Added & Functioning
Progress Bar added sucessfully. Tested and it works with the current layout (3 speed modes).

If there are issues witht the filling percentage (ProgressBarLinear.current), the problem might be with the GameObjects. Make sure the scale for the linear progress bar is 1,1,1.

Phantom arm replaced by avatar arm in the test run (last elbow rotation in each speed mode).